feat(errors): show foother/header if exception inside page
