+++ src/App.js
@@ -20,7 +20,7 @@ import MobileNavbar from './components/MobileNavbar/MobileNavbar'
 import Navbar from './components/Navbar/Navbar'
 import withTracker from './withTracker'
 import withIntercom from './withIntercom'
-import ErrorBoundary from './ErrorBoundary'
+import ErrorBoundary from './components/ErrorContent/ErrorBoundary'
 import PageLoader from './components/Loader/PageLoader'
 import Footer from './components/Footer'
 import GDPRPage from './pages/GDPRPage/GDPRPage'
@@ -64,11 +64,6 @@ const LoadableHistoricalBalancePage = Loadable({
 
 const LoadableMarketingPage = Loadable({
   loader: () => import('./pages/Marketing/MarketingPage'),
-  loading: () => <PageLoader />
-})
-
-const LoadableErrorPage = Loadable({
-  loader: () => import('./pages/ErrorPage/ErrorPage'),
   loading: () => <PageLoader />
 })
 
@@ -260,239 +255,243 @@ export const App = ({
       )}
       <GdprRedirector pathname={pathname} />
       {isDesktop && <UrlModals />}
-      <Switch>
-        <Route path={SHARE_PATH} component={PageLoader} />
-        {['erc20', 'all', 'list', 'screener'].map(name => (
-          <Route
-            exact
-            key={name}
-            path={`/assets/${name}`}
-            render={props => {
-              if (isDesktop) {
+
+      <ErrorBoundary>
+        <Switch>
+          <Route path={SHARE_PATH} component={PageLoader} />
+          {['erc20', 'all', 'list', 'screener'].map(name => (
+            <Route
+              exact
+              key={name}
+              path={`/assets/${name}`}
+              render={props => {
+                if (isDesktop) {
+                  return (
+                    <WatchlistPage
+                      type={name}
+                      isLoggedIn={isLoggedIn}
+                      preload={() => LoadableDetailedPage.preload()}
+                      {...props}
+                    />
+                  )
+                }
                 return (
-                  <WatchlistPage
+                  <LoadableAssetsMobilePage
                     type={name}
                     isLoggedIn={isLoggedIn}
-                    preload={() => LoadableDetailedPage.preload()}
                     {...props}
                   />
                 )
-              }
-              return (
-                <LoadableAssetsMobilePage
-                  type={name}
-                  isLoggedIn={isLoggedIn}
-                  {...props}
-                />
+              }}
+            />
+          ))}
+          <Route exact path='/pricing' component={LoadablePricingPage} />
+          <Route
+            exact
+            path={PATHS.GDPR}
+            render={props => <GDPRPage {...props} isDesktop={isDesktop} />}
+          />
+          <Route
+            exact
+            path={PATHS.CREATE_ACCOUNT}
+            render={props => (
+              <CreateAccountFreeTrial {...props} isLoggedIn={isLoggedIn} />
+            )}
+          />
+          <Route exact path='/assets' component={LoadableWatchlistsPage} />
+          <Route
+            exact
+            path='/watchlists'
+            render={props =>
+              isDesktop ? (
+                <Redirect from='/watchlists' to='/assets' />
+              ) : (
+                <LoadableWatchlistsMobilePage {...props} />
               )
+            }
+          />
+          <Route
+            exact
+            path='/unsubscribe'
+            component={LoadableUnsubscribePage}
+          />
+          <Route
+            path={PATHS.FEED}
+            render={props => <LoadableFeedPage {...props} />}
+          />
+          <Route
+            exact
+            path='/search'
+            render={props => {
+              if (isDesktop) {
+                return <Redirect to='/' />
+              }
+              return <LoadableSearchMobilePage {...props} />
             }}
           />
-        ))}
-        <Route exact path='/pricing' component={LoadablePricingPage} />
-        <Route
-          exact
-          path={PATHS.GDPR}
-          render={props => <GDPRPage {...props} isDesktop={isDesktop} />}
-        />
-        <Route
-          exact
-          path={PATHS.CREATE_ACCOUNT}
-          render={props => (
-            <CreateAccountFreeTrial {...props} isLoggedIn={isLoggedIn} />
-          )}
-        />
-        <Route exact path='/assets' component={LoadableWatchlistsPage} />
-        <Route
-          exact
-          path='/watchlists'
-          render={props =>
-            isDesktop ? (
-              <Redirect from='/watchlists' to='/assets' />
-            ) : (
-              <LoadableWatchlistsMobilePage {...props} />
-            )
-          }
-        />
-        <Route exact path='/unsubscribe' component={LoadableUnsubscribePage} />
-        <Route
-          path={PATHS.FEED}
-          render={props => <LoadableFeedPage {...props} />}
-        />
-        <Route
-          exact
-          path='/search'
-          render={props => {
-            if (isDesktop) {
-              return <Redirect to='/' />
+          <Route exact path='/roadmap' component={Roadmap} />
+          <Route
+            exact
+            path='/labs/balance'
+            render={({ history }) => (
+              <LoadableHistoricalBalancePage
+                history={history}
+                isDesktop={isDesktop}
+              />
+            )}
+          />
+          <Route
+            exact
+            path='/projects/:slug'
+            render={props =>
+              isDesktop ? (
+                <LoadableDetailedPage isDesktop={isDesktop} {...props} />
+              ) : (
+                <LoadableMobileDetailedPage {...props} />
+              )
             }
-            return <LoadableSearchMobilePage {...props} />
-          }}
-        />
-        <Route exact path='/roadmap' component={Roadmap} />
-        <Route
-          exact
-          path='/labs/balance'
-          render={({ history }) => (
-            <LoadableHistoricalBalancePage
-              history={history}
-              isDesktop={isDesktop}
-            />
-          )}
-        />
-        <Route
-          exact
-          path='/projects/:slug'
-          render={props =>
-            isDesktop ? (
-              <LoadableDetailedPage isDesktop={isDesktop} {...props} />
-            ) : (
-              <LoadableMobileDetailedPage {...props} />
-            )
-          }
-        />
-        <Route exact path='/labs/trends' component={LoadableTrendsLabsPage} />
-        <Route exact path='/labs' component={LoadableLabsPage} />
-        <Redirect from='/trends' to='/labs/trends' />
-        <Route
-          exact
-          path={['/labs/trends/explore/:word', '/labs/trends/explore/']}
-          render={props => (
-            <LoadableTrendsExplorePage isDesktop={isDesktop} {...props} />
-          )}
-        />
-        <Route
-          path='/sonar'
-          render={props => (
-            <LoadableSonarFeedPage
-              isDesktop={isDesktop}
-              isLoggedIn={isLoggedIn}
-              {...props}
-            />
-          )}
-        />
-        <Route path='/logout' component={LogoutPage} />
-        <Route
-          exact
-          path='/account'
-          render={props => (
-            <LoadableAccountPage
-              {...props}
-              isUserLoading={isUserLoading}
-              isLoggedIn={isLoggedIn}
-            />
-          )}
-        />
-        <Redirect from='/ethereum-spent' to='/projects/ethereum' />
-        <Route exact path='/privacy-policy' component={PrivacyPolicyPage} />
-        <Route path='/email_login' component={EmailLoginVerification} />
-        <Route path='/verify_email' component={EmailLoginVerification} />
-        {ExternalRoutes.map(links => {
-          return links.routes.map(name => (
-            <Route
-              key={name}
-              path={`/${name}`}
-              exact
-              render={() => <ExternalRedirect to={links.to} />}
-            />
-          ))
-        })}
-        <Route
-          path='/consent'
-          render={props => (
-            <ExternalRedirect
-              to={`${getConsentUrl()}/consent${props.location.search}`}
-            />
-          )}
-        />{' '}
-        <Route
-          path={['/profile/:id', '/profile']}
-          render={props => (
-            <LoadableProfilePage
-              isDesktop={isDesktop}
-              location={props.location}
-              {...props}
-            />
-          )}
-        />
-        <Route
-          path={PATHS.LOGIN}
-          render={props => (
-            <LoadableLoginPage
-              isLoggedIn={isLoggedIn}
-              token={token}
-              {...props}
-            />
-          )}
-        />
-        <Route
-          path={PATHS.PRO_METRICS}
-          render={props => (
-            <LoadableProMetricsPage isLoggedIn={isLoggedIn} {...props} />
-          )}
-        />
-        {!isDesktop && <Redirect from={PATHS.STUDIO} to='/assets' />}
-        <Route
-          path={PATHS.STUDIO}
-          render={props => (
-            <LoadableChartPage
-              classes={{ wrapper: styles.chart }}
-              isLoggedIn={isLoggedIn}
-              {...props}
-            />
-          )}
-        />
-        <Route
-          path={PATHS.STABLECOINS}
-          render={props => (
-            <LoadableStablecoinsPage isDesktop={isDesktop} {...props} />
-          )}
-        />
-        <Route
-          path={PATHS.UNISWAP_PROTOCOL}
-          render={props => (
-            <LoadableUniswapProtocolPage isDesktop={isDesktop} {...props} />
-          )}
-        />
-        <Route
-          path={PATHS.DEXS}
-          render={props => (
-            <LoadableDexsPage isDesktop={isDesktop} {...props} />
-          )}
-        />
-        <Route
-          path={PATHS.SHEETS_TEMPLATES}
-          render={props => (
-            <LoadableSheetsTemplatePage
-              isLoggedIn={isLoggedIn}
-              isDesktop={isDesktop}
-              {...props}
-            />
-          )}
-        />
-        <Route
-          path={PATHS.CHARTS}
-          render={props => (
-            <LoadableChartPage
-              classes={{ wrapper: styles.chart }}
-              isLoggedIn={isLoggedIn}
-              {...props}
-            />
-          )}
-        />
-        <Route
-          path={PATHS.ERROR}
-          render={props => <LoadableErrorPage {...props} />}
-        />
-        {!isDesktop && <Redirect from={PATHS.INDEX} to='/assets' />}
-        <Route
-          path={PATHS.INDEX}
-          render={props => (
-            <LoadableMarketingPage isLoggedIn={isLoggedIn} {...props} />
-          )}
-        />
-      </Switch>
-      <NotificationStack />
-      <CookiePopup />
+          />
+          <Route exact path='/labs/trends' component={LoadableTrendsLabsPage} />
+          <Route exact path='/labs' component={LoadableLabsPage} />
+          <Redirect from='/trends' to='/labs/trends' />
+          <Route
+            exact
+            path={['/labs/trends/explore/:word', '/labs/trends/explore/']}
+            render={props => (
+              <LoadableTrendsExplorePage isDesktop={isDesktop} {...props} />
+            )}
+          />
+          <Route
+            path='/sonar'
+            render={props => (
+              <LoadableSonarFeedPage
+                isDesktop={isDesktop}
+                isLoggedIn={isLoggedIn}
+                {...props}
+              />
+            )}
+          />
+          <Route path='/logout' component={LogoutPage} />
+          <Route
+            exact
+            path='/account'
+            render={props => (
+              <LoadableAccountPage
+                {...props}
+                isUserLoading={isUserLoading}
+                isLoggedIn={isLoggedIn}
+              />
+            )}
+          />
+          <Redirect from='/ethereum-spent' to='/projects/ethereum' />
+          <Route exact path='/privacy-policy' component={PrivacyPolicyPage} />
+          <Route path='/email_login' component={EmailLoginVerification} />
+          <Route path='/verify_email' component={EmailLoginVerification} />
+          {ExternalRoutes.map(links => {
+            return links.routes.map(name => (
+              <Route
+                key={name}
+                path={`/${name}`}
+                exact
+                render={() => <ExternalRedirect to={links.to} />}
+              />
+            ))
+          })}
+          <Route
+            path='/consent'
+            render={props => (
+              <ExternalRedirect
+                to={`${getConsentUrl()}/consent${props.location.search}`}
+              />
+            )}
+          />{' '}
+          <Route
+            path={['/profile/:id', '/profile']}
+            render={props => (
+              <LoadableProfilePage
+                isDesktop={isDesktop}
+                location={props.location}
+                {...props}
+              />
+            )}
+          />
+          <Route
+            path={PATHS.LOGIN}
+            render={props => (
+              <LoadableLoginPage
+                isLoggedIn={isLoggedIn}
+                token={token}
+                {...props}
+              />
+            )}
+          />
+          <Route
+            path={PATHS.PRO_METRICS}
+            render={props => (
+              <LoadableProMetricsPage isLoggedIn={isLoggedIn} {...props} />
+            )}
+          />
+          {!isDesktop && <Redirect from={PATHS.STUDIO} to='/assets' />}
+          <Route
+            path={PATHS.STUDIO}
+            render={props => (
+              <LoadableChartPage
+                classes={{ wrapper: styles.chart }}
+                isLoggedIn={isLoggedIn}
+                {...props}
+              />
+            )}
+          />
+          <Route
+            path={PATHS.STABLECOINS}
+            render={props => (
+              <LoadableStablecoinsPage isDesktop={isDesktop} {...props} />
+            )}
+          />
+          <Route
+            path={PATHS.UNISWAP_PROTOCOL}
+            render={props => (
+              <LoadableUniswapProtocolPage isDesktop={isDesktop} {...props} />
+            )}
+          />
+          <Route
+            path={PATHS.DEXS}
+            render={props => (
+              <LoadableDexsPage isDesktop={isDesktop} {...props} />
+            )}
+          />
+          <Route
+            path={PATHS.SHEETS_TEMPLATES}
+            render={props => (
+              <LoadableSheetsTemplatePage
+                isLoggedIn={isLoggedIn}
+                isDesktop={isDesktop}
+                {...props}
+              />
+            )}
+          />
+          <Route
+            path={PATHS.CHARTS}
+            render={props => (
+              <LoadableChartPage
+                classes={{ wrapper: styles.chart }}
+                isLoggedIn={isLoggedIn}
+                {...props}
+              />
+            )}
+          />
+          {!isDesktop && <Redirect from={PATHS.INDEX} to='/assets' />}
+          <Route
+            path={PATHS.INDEX}
+            render={props => (
+              <LoadableMarketingPage isLoggedIn={isLoggedIn} {...props} />
+            )}
+          />
+        </Switch>
+        <NotificationStack />
+        <CookiePopup />
+      </ErrorBoundary>
+
       {isDesktop && showFooter && (
         <Footer
           classes={{
