+++ src/paths.js
@@ -12,6 +12,5 @@ export const PATHS = {
   STABLECOINS: '/stablecoins',
   UNISWAP_PROTOCOL: '/uniswap-protocol',
   SHEETS_TEMPLATES: '/sheets/daa-vs-price',
-  DEXS: '/decentralized-exchanges',
-  ERROR: '/error'
+  DEXS: '/decentralized-exchanges'
 }
